fix(check): respect Nuxt srcDir fallback aliases
