+++ crates/vize/src/commands/check/nuxt/tests.rs
@@ -16,6 +16,7 @@ use super::stubs::declared_name;
 
 mod build_dir;
 mod config_modules;
+mod path_aliases;
 
 #[test]
 fn parses_module_export_lines() {
@@ -600,103 +601,6 @@ export default defineNuxtConfig({
                 .any(|stub| stub.starts_with(absent)),
             "commented-out module must not inject {absent:?}, got: {:#?}",
             options.auto_import_stubs
-        );
-    }
-
-    let _ = std::fs::remove_dir_all(&project_root);
-}
-
-#[test]
-fn path_aliases_come_from_generated_nuxt_tsconfig_when_present() {
-    let project_root = unique_case_dir("nuxt-generated-tsconfig-aliases");
-    let _ = std::fs::remove_dir_all(&project_root);
-    std::fs::create_dir_all(project_root.join(".nuxt")).unwrap();
-    std::fs::create_dir_all(project_root.join("app")).unwrap();
-    std::fs::write(project_root.join("nuxt.config.ts"), "export default {}").unwrap();
-    // `.nuxt/tsconfig.json` is JSON-with-comments and carries the project's REAL
-    // `compilerOptions.paths`. Targets are relative to `.nuxt/`. The custom
-    // `@features/*` alias would never be produced by the hardcoded guesses.
-    std::fs::write(
-        project_root.join(".nuxt/tsconfig.json"),
-        r##"{
-  // generated by nuxi prepare
-  "compilerOptions": {
-    "paths": {
-      "~/*": ["../app/*"],
-      "@features/*": ["../app/features/*"],
-      "#shared/*": ["../shared/*"],
-    },
-  },
-}
-"##,
-    )
-    .unwrap();
-
-    let mut options = VirtualTsOptions::default();
-    let aliases = detect_nuxt_auto_imports(&mut options, &project_root);
-
-    // Aliases come verbatim from the generated config, rebased to project root.
-    assert!(
-        aliases.iter().any(|alias| {
-            alias.pattern.as_str() == "~/*"
-                && alias
-                    .targets
-                    .iter()
-                    .any(|target| target.as_str() == "app/*")
-        }),
-        "expected rebased ~/* alias, got: {aliases:#?}"
-    );
-    assert!(
-        aliases.iter().any(|alias| {
-            alias.pattern.as_str() == "@features/*"
-                && alias
-                    .targets
-                    .iter()
-                    .any(|target| target.as_str() == "app/features/*")
-        }),
-        "expected custom @features/* alias from generated tsconfig, got: {aliases:#?}"
-    );
-    assert!(
-        aliases.iter().any(|alias| {
-            alias.pattern.as_str() == "#shared/*"
-                && alias
-                    .targets
-                    .iter()
-                    .any(|target| target.as_str() == "shared/*")
-        }),
-        "expected #shared/* alias from generated tsconfig, got: {aliases:#?}"
-    );
-    // Hardcoded guesses absent from the generated config must NOT be synthesized.
-    assert!(
-        !aliases.iter().any(|alias| alias.pattern.as_str() == "@/*"),
-        "guessed @/* alias must not be injected when generated tsconfig is present: {aliases:#?}"
-    );
-
-    let _ = std::fs::remove_dir_all(&project_root);
-}
-
-#[test]
-fn path_aliases_fall_back_to_guesses_without_generated_nuxt_tsconfig() {
-    let project_root = unique_case_dir("nuxt-no-generated-tsconfig-aliases");
-    let _ = std::fs::remove_dir_all(&project_root);
-    std::fs::create_dir_all(project_root.join("app")).unwrap();
-    std::fs::create_dir_all(project_root.join("shared")).unwrap();
-    std::fs::write(project_root.join("nuxt.config.ts"), "export default {}").unwrap();
-
-    let mut options = VirtualTsOptions::default();
-    let aliases = detect_nuxt_auto_imports(&mut options, &project_root);
-
-    for (pattern, target) in [
-        ("~/*", "app/*"),
-        ("@/*", "app/*"),
-        ("#shared/*", "shared/*"),
-    ] {
-        assert!(
-            aliases.iter().any(|alias| {
-                alias.pattern.as_str() == pattern
-                    && alias.targets.iter().any(|t| t.as_str() == target)
-            }),
-            "expected hardcoded fallback alias {pattern:?} -> {target:?}, got: {aliases:#?}"
         );
     }
