+++ crates/vize/src/commands/check/nuxt/tests/path_aliases.rs
@@ -1,0 +1,110 @@
+use vize_canon::virtual_ts::VirtualTsOptions;
+
+use super::super::detect_nuxt_auto_imports;
+use super::unique_case_dir;
+
+#[test]
+fn path_aliases_come_from_generated_nuxt_tsconfig_when_present() {
+    let project_root = unique_case_dir("nuxt-generated-tsconfig-aliases");
+    let _ = std::fs::remove_dir_all(&project_root);
+    std::fs::create_dir_all(project_root.join(".nuxt")).unwrap();
+    std::fs::create_dir_all(project_root.join("app")).unwrap();
+    std::fs::write(project_root.join("nuxt.config.ts"), "export default {}").unwrap();
+    // `.nuxt/tsconfig.json` is JSON-with-comments and carries the project's REAL
+    // `compilerOptions.paths`. Targets are relative to `.nuxt/`. The custom
+    // `@features/*` alias would never be produced by the hardcoded guesses.
+    std::fs::write(
+        project_root.join(".nuxt/tsconfig.json"),
+        r##"{
+  // generated by nuxi prepare
+  "compilerOptions": {
+    "paths": {
+      "~/*": ["../app/*"],
+      "@features/*": ["../app/features/*"],
+      "#shared/*": ["../shared/*"],
+    },
+  },
+}
+"##,
+    )
+    .unwrap();
+
+    let mut options = VirtualTsOptions::default();
+    let aliases = detect_nuxt_auto_imports(&mut options, &project_root);
+
+    assert_alias(&aliases, "~/*", "app/*");
+    assert_alias(&aliases, "@features/*", "app/features/*");
+    assert_alias(&aliases, "#shared/*", "shared/*");
+    assert!(
+        !aliases.iter().any(|alias| alias.pattern.as_str() == "@/*"),
+        "guessed @/* alias must not be injected when generated tsconfig is present: {aliases:#?}"
+    );
+
+    let _ = std::fs::remove_dir_all(&project_root);
+}
+
+#[test]
+fn path_aliases_fall_back_to_guesses_without_generated_nuxt_tsconfig() {
+    let project_root = unique_case_dir("nuxt-no-generated-tsconfig-aliases");
+    let _ = std::fs::remove_dir_all(&project_root);
+    std::fs::create_dir_all(project_root.join("app")).unwrap();
+    std::fs::create_dir_all(project_root.join("shared")).unwrap();
+    std::fs::write(project_root.join("nuxt.config.ts"), "export default {}").unwrap();
+
+    let mut options = VirtualTsOptions::default();
+    let aliases = detect_nuxt_auto_imports(&mut options, &project_root);
+
+    assert_alias(&aliases, "~/*", "app/*");
+    assert_alias(&aliases, "@/*", "app/*");
+    assert_alias(&aliases, "#shared/*", "shared/*");
+
+    let _ = std::fs::remove_dir_all(&project_root);
+}
+
+#[test]
+fn path_aliases_use_static_src_dir_without_generated_nuxt_tsconfig() {
+    let project_root = unique_case_dir("nuxt-src-dir-tsconfig-aliases");
+    let _ = std::fs::remove_dir_all(&project_root);
+    std::fs::create_dir_all(project_root.join("src/pages")).unwrap();
+    std::fs::write(
+        project_root.join("nuxt.config.ts"),
+        r#"export default defineNuxtConfig({ srcDir: "src" })"#,
+    )
+    .unwrap();
+
+    let mut options = VirtualTsOptions::default();
+    let aliases = detect_nuxt_auto_imports(&mut options, &project_root);
+
+    assert_alias(&aliases, "~/*", "src/*");
+    assert_alias(&aliases, "@/*", "src/*");
+    assert_no_alias(&aliases, "~/*", "*");
+    assert_no_alias(&aliases, "@/*", "*");
+
+    let _ = std::fs::remove_dir_all(&project_root);
+}
+
+fn assert_no_alias(aliases: &[super::super::NuxtPathAlias], pattern: &str, target: &str) {
+    assert!(
+        !aliases.iter().any(|alias| {
+            alias.pattern.as_str() == pattern
+                && alias
+                    .targets
+                    .iter()
+                    .any(|candidate| candidate.as_str() == target)
+        }),
+        "alias {pattern:?} must not point at {target:?}, got: {aliases:#?}"
+    );
+}
+
+fn assert_alias(aliases: &[super::super::NuxtPathAlias], pattern: &str, target: &str) {
+    assert!(
+        aliases.iter().any(|alias| {
+            alias.pattern.as_str() == pattern
+                && alias
+                    .targets
+                    .iter()
+                    .any(|candidate| candidate.as_str() == target)
+        }),
+        "expected alias {pattern:?} -> {target:?}, got: {aliases:#?}"
+    );
+}
